requirements.txt
Django>=4.2,<5.1
djangorestframework>=3.14
celery>=5.3
redis>=5.0
requests>=2.31
pydub>=0.25.1
python-dotenv>=1.0.1
djangorestframework-simplejwt>=5.3.1
channels>=4.1
channels-redis>=4.2